# Conflicts:
#	src/main/java/com/backendteam5/finalproject/controller/HomeController.java
#	src/main/java/com/backendteam5/finalproject/entity/Courier.java
#	src/main/java/com/backendteam5/finalproject/repository/CourierRepository.java
#	src/main/java/com/backendteam5/finalproject/service/AdminService.java
#	src/main/java/com/backendteam5/finalproject/service/CourierService.java
#	src/main/resources/static/js/main.js
#	src/main/resources/templates/index1.html
#	src/main/resources/templates/index2.html

+++ src/main/java/com/backendteam5/finalproject/FinalprojectApplication.java
@@ -7,7 +7,6 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -23,7 +22,7 @@ public class FinalprojectApplication implements CommandLineRunner {
     @Override
     public void run(String... args) throws Exception{
         List<CourierInfo> courierInfoList = new LinkedList<>();
-        for(long i = 1L; i <= 20L; i++){
+        for(int i = 1; i <= 20; i++){
             courierInfoList.add(new CourierInfo("A", i, "구로구 "+ i +"지역"));
         }
         courierInfoRepository.saveAll(courierInfoList);
